Refactor: KakaoService에서 response 부분 수정
