+++ src/main/java/com/kwcapstone/Controller/MemberController.java
@@ -5,7 +5,7 @@ import com.kwcapstone.Domain.Dto.Request.EmailDuplicationDto;
 import com.kwcapstone.Domain.Dto.Request.EmailRequestDto;
 import com.kwcapstone.Domain.Dto.Request.MemberRequestDto;
 import com.kwcapstone.Domain.Entity.Member;
-import com.kwcapstone.GoogleLogin.Auth.SessionUser;
+//import com.kwcapstone.GoogleLogin.Auth.SessionUser;
 import com.kwcapstone.Repository.MemberRepository;
 import com.kwcapstone.Service.MemberService;
 import jakarta.servlet.http.HttpSession;
@@ -47,6 +47,7 @@ public class MemberController {
         return new BaseResponse(HttpStatus.OK.value(), "이메일 인증이 완료되었습니다.");
     }
 
+    /*
     // 약관동의
     @PostMapping("/auth/agree")
     public BaseResponse agree() {
@@ -65,6 +66,6 @@ public class MemberController {
         httpSession.removeAttribute("tempMember");
 
         return new BaseResponse(HttpStatus.OK.value(), "회원가입이 완료되었습니다.");
-    }
+    }*/
 
 }
