+++ src/main/java/com/kwcapstone/Kakao/Dto/OAuthToken.java
@@ -1,4 +1,4 @@
-package com.kwcapstone.Kakao;
+package com.kwcapstone.Kakao.Dto;
 
 import lombok.Data;
 
@@ -6,9 +6,9 @@ import lombok.Data;
 public class OAuthToken {
     private String access_token; //엑세스
     private String refresh_token; //리프레쉬
-    private Long id_token; //토큰 Id값
+    //private Long id_token; //토큰 Id값
     private Integer expires_in;//엑세스 만료시간
-    private Integer refresh_expires_in;//리프레쉬 토큰 만료 시간
+    private Integer refresh_token_expires_in;//리프레쉬 토큰 만료 시간
     private String token_type; //토큰타입
     private String scope; //인증된 사용자의 정보 조회 권한 범위
 }
