+++ src/main/java/com/kwcapstone/Kakao/Service/KaKaoProvider.java
@@ -1,8 +1,11 @@
-package com.kwcapstone.Kakao;
+package com.kwcapstone.Kakao.Service;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.kwcapstone.Kakao.Dto.KaKaoProfile;
+import com.kwcapstone.Kakao.Dto.OAuthToken;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.*;
 import org.springframework.stereotype.Component;
@@ -11,6 +14,7 @@ import org.springframework.util.MultiValueMap;
 import org.springframework.web.client.RestTemplate;
 import org.springframework.web.server.ResponseStatusException;
 
+@Slf4j
 @Component
 @RequiredArgsConstructor
 public class KaKaoProvider {
@@ -21,17 +25,23 @@ public class KaKaoProvider {
     @Value("${KAKAO_REDIRECT_URL}")
     private String redirectUrl;
 
+    @Value("${KAKAO_CLIENT_SECRET}")
+    private String clientSecret;
+
     //code(인가코드)로 accessToken 요청하기
     // (해당 accessToken은 카카오에서 제공해주는 token)
     //보안을 위해 accessToken을 새로 서버쪽에서 발급할거임.
     public OAuthToken requestToken(String code){
+        System.out.println("clientId" + clientId);
+        System.out.println("redirectUrl" + redirectUrl);
+
         //Restemplate 새로 생성(Http 요청을 보내기 위함)
         RestTemplate restTemplate = new RestTemplate();
 
         //Http 요청의 header에 방식(Json이 아닌 URL 인코딩된 폼 데이터)
         HttpHeaders headers = new HttpHeaders();
-        headers.add("Content-Type",
-                "application/x-www-form-urlencoded;charset=UTF-8");
+        headers.add("Content-type",
+                "application/x-www-form-urlencoded;charset=utf-8");
 
         //넘길 정보 담기
         MultiValueMap<String, String> params
@@ -40,16 +50,21 @@ public class KaKaoProvider {
         params.add("client_id", clientId);
         params.add("redirect_uri", redirectUrl);
         params.add("code", code);
+        params.add("client_secret", clientSecret);
+
+        System.out.println("params" + params);
 
         //요청할 객체 생성
-        HttpEntity<MultiValueMap<String,String>> kakaoTokenRequest
+        HttpEntity<MultiValueMap<String, String>> kakaoTokenRequest
                 = new HttpEntity<>(params, headers);
 
         //카카오 OAuth 토큰 요청하기
         ResponseEntity<String> response
                 = restTemplate.exchange(
-                        "https://kauth.kakao.com/oauth/token",
-                HttpMethod.POST, kakaoTokenRequest, String.class);
+                        "https://kauth.kakao.com/oauth/token", HttpMethod.POST, kakaoTokenRequest, String.class);
+
+        System.out.println(response.getStatusCode());
+        System.out.println(response.getBody());
 
         //응답데이터는 OAuthToken로 변환해야함(이를 위해 객체 매핑할 도구)
         ObjectMapper objectMapper = new ObjectMapper();
